Migrate guess_from_pecl to rust
