+++ src/providers/mod.rs
@@ -12,6 +12,7 @@ pub mod launchpad;
 pub mod maven;
 pub mod meson;
 pub mod metainfo;
+pub mod nuspec;
 #[cfg(feature = "opam")]
 pub mod ocaml;
 pub mod package_json;
