+++ upstream_ontologist/_upstream_ontologist.pyi
@@ -100,3 +100,4 @@ def fixup_broken_git_details(url: str, branch: str | None = None, subpath: str |
 def guess_upstream_info(
     path: str, trust_package: bool = False
 ) -> list[UpstreamDatum]: ...
+def guess_from_pecl_package(package: str) -> Iterator[tuple[str, UpstreamDatum]]: ...
